+++ services/horizon/CHANGELOG.md
@@ -9,6 +9,7 @@ bumps.  A breaking change will get clearly notified in this log.
 ## v0.18.2
 
 * Added configuration toggle to return JSON response with unindented JSON. This improves rendering and request rate performance.
+* Added configuration toggle to disable non-paginating HAL `"_links"` in JSON response. This improves rendering and request rate performance.
 
 ## v0.18.1